+++ lib/maze/driver/appium.rb
@@ -99,7 +99,7 @@ module Maze
         end
       rescue Selenium::WebDriver::Error::ServerError => e
         # Assume the remote appium session has stopped, so crash out of the session
-        fail_driver(e.message)
+        fail_driver(e)
         raise e
       else
         true
@@ -110,7 +110,7 @@ module Maze
         super
       rescue Selenium::WebDriver::Error::ServerError => e
         # Assume the remote appium session has stopped, so crash out of the session
-        fail_driver(e.message)
+        fail_driver(e)
         raise e
       end
 
@@ -119,7 +119,7 @@ module Maze
         super
       rescue Selenium::WebDriver::Error::ServerError => e
         # Assume the remote appium session has stopped, so crash out of the session
-        fail_driver(e.message)
+        fail_driver(e)
         raise e
       end
 
@@ -130,7 +130,7 @@ module Maze
         end
       rescue Selenium::WebDriver::Error::ServerError => e
         # Assume the remote appium session has stopped, so crash out of the session
-        fail_driver(e.message)
+        fail_driver(e)
         raise e
       end
 
@@ -144,7 +144,7 @@ module Maze
         end
       rescue Selenium::WebDriver::Error::ServerError => e
         # Assume the remote appium session has stopped, so crash out of the session
-        fail_driver(e.message)
+        fail_driver(e)
         raise e
       end
 
@@ -162,7 +162,7 @@ module Maze
         false
       rescue Selenium::WebDriver::Error::ServerError => e
         # Assume the remote appium session has stopped, so crash out of the session
-        fail_driver(e.message)
+        fail_driver(e)
         raise e
       end